+++ venv/DataVisualization/boxes.py
@@ -6,7 +6,6 @@ import click
 
 # Check '18 - '19 different statistical categories for the teams
 # Dataframes - dfs, games.csv is interesting enough in combination with the others that have already been used
-games_details = pd.read_csv('../Data/games_details.csv')
 games = pd.read_csv('../Data/games.csv')
 teams = pd.read_csv('../Data/teams.csv')